# smo/docs/openapi/sa-smos.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntentNotification": {
        "description": "What Intent Service pushes to a handling function on CreateIntent.",
        "properties": {
          "intentId": {
            "format": "uuid",
            "title": "Intentid",
            "type": "string"
          }
        },
        "required": [
          "intentId"
        ],
        "title": "IntentNotification",
        "type": "object"
      },
      "RegistrationBody": {
        "additionalProperties": false,
        "properties": {
          "cmTargets": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmtargets"
          },
          "intentHandlingScope": {
            "default": [
              "RAN"
            ],
            "items": {
              "type": "string"
            },
            "title": "Intenthandlingscope",
            "type": "array"
          }
        },
        "title": "RegistrationBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "r1BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SA SMOS",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Liveness probe, kept as an alias of `/live`. The GUI BFF's `GET /modules/status` fans out to\n`/<module>/health` through R1 Termination for every module in parallel, and DME supervises\nproducers on it.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Check",
        "tags": [
          "health"
        ]
      }
    },
    "/live": {
      "get": {
        "description": "Liveness probe: the process is up. Depends on nothing else.",
        "operationId": "live_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live",
        "tags": [
          "health"
        ]
      }
    },
    "/monitors": {
      "get": {
        "description": "List read over AssuranceMonitor \u2014 previously write-only, so the\nthresholds being evaluated were invisible to an operator.",
        "operationId": "list_assurance_monitors_monitors_get",
        "parameters": [
          {
            "description": "Max rows to return (1-500).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max rows to return (1-500).",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip before the first one returned.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the first one returned.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assurance Monitors"
      },
      "post": {
        "description": "At most one target: a service order (NF deployment remediation), a\nmodel coordination group (retrain), or a rApp instance (remediation\nthrough rApp Management, including ROLLBACK).",
        "operationId": "register_assurance_monitor_monitors_post",
        "parameters": [
          {
            "in": "query",
            "name": "target_order_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Order Id"
            }
          },
          {
            "in": "query",
            "name": "target_coordination_group_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Coordination Group Id"
            }
          },
          {
            "in": "query",
            "name": "target_rapp_instance_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Rapp Instance Id"
            }
          },
          {
            "in": "query",
            "name": "analytics_subscription_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analytics Subscription Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Thresholds",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Assurance Monitor"
      }
    },
    "/monitors/{monitor_id}": {
      "get": {
        "operationId": "get_assurance_monitor_monitors__monitor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "monitor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Monitor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assurance Monitor"
      }
    },
    "/monitors/{monitor_id}/escalate": {
      "post": {
        "operationId": "escalate_to_operator_monitors__monitor_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "monitor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Monitor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": true,
            "schema": {
              "title": "Reason",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Escalate To Operator"
      }
    },
    "/monitors/{monitor_id}/evaluate": {
      "post": {
        "description": "Periodic, aligned with RAN Analytics' subscription cadence\n(D-NFR-PERF-SA-1, unchanged from v1.3 \u2014 this was already specified,\nnot a gap this LLD pass needed to close).",
        "operationId": "evaluate_thresholds_monitors__monitor_id__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "monitor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Monitor Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Current Metrics",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Thresholds"
      }
    },
    "/monitors/{monitor_id}/remedial-actions": {
      "post": {
        "description": "ExecuteRemedialAction \u2014 gated by autoExecutionScopeConfig, admin-only\nto change (REQ-CNFG-ADM pattern, unchanged). Dispatch per SO/SA SMOS\nLLD section 2.1: CONFIG_CHANGE, SCALE, and RECONNECT are resolved;\nROLLBACK is resolved for a rApp-instance-scoped monitor and refused\n(409 ROLLBACK_HISTORY_UNAVAILABLE) for any other (see the module\ndocstring). A coordination-group-scoped monitor bypasses all four\nactionType branches below \u2014 see the module docstring's\nMLModelCoordinationGroup convergence note.\n\nA rApp Management refusal (nothing left to roll back, the instance not\nRUNNING, the earlier package gone) is not an error here: the action is\nrecorded ESCALATED, with rApp Management's reason in `detail`. A\nRESOLVED rollback has started an upgrade back; `result` names the\nreplacement instance, which rApp Management resolves like any upgrade.",
        "operationId": "execute_remedial_action_monitors__monitor_id__remedial_actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "monitor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Monitor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action_type",
            "required": true,
            "schema": {
              "title": "Action Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "requester_is_admin",
            "required": false,
            "schema": {
              "default": false,
              "title": "Requester Is Admin",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Remedial Action"
      }
    },
    "/o1-cm-handler/enactments": {
      "get": {
        "operationId": "list_enactments_o1_cm_handler_enactments_get",
        "parameters": [
          {
            "in": "query",
            "name": "intent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Intent Id"
            }
          },
          {
            "description": "Max rows to return (1-500).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max rows to return (1-500).",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip before the first one returned.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the first one returned.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Enactments"
      }
    },
    "/o1-cm-handler/intents": {
      "post": {
        "description": "Intent Service's push for a new Intent addressed to this handler.",
        "operationId": "enact_intent_o1_cm_handler_intents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentNotification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enact Intent"
      }
    },
    "/o1-cm-handler/registration": {
      "delete": {
        "operationId": "deregister_o1_cm_handler_o1_cm_handler_registration_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Deregister O1 Cm Handler"
      },
      "post": {
        "description": "Registers SA SMOS with Intent Service as the O1-CM intent handling\nfunction. Idempotent: an existing registration is replaced.",
        "operationId": "register_o1_cm_handler_o1_cm_handler_registration_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationBody",
                "default": {
                  "intentHandlingScope": [
                    "RAN"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register O1 Cm Handler"
      }
    },
    "/ready": {
      "get": {
        "description": "Readiness probe: 200 when the database and SME (where this module uses them) answer, else 503\nnaming the failing check. Take the replica out of rotation on 503; do not restart it.",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    },
    "/remedial-actions": {
      "get": {
        "description": "Every RemedialAction row, optionally per-monitor or per-outcome \u2014\n`outcome=ESCALATED` is the operator's escalation queue (the GUI\ndashboard's SA SMOS tile).",
        "operationId": "list_remedial_actions_remedial_actions_get",
        "parameters": [
          {
            "in": "query",
            "name": "monitor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Monitor Id"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "description": "Max rows to return (1-500).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max rows to return (1-500).",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip before the first one returned.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the first one returned.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Remedial Actions"
      }
    }
  },
  "security": [
    {
      "r1BearerAuth": []
    }
  ]
}